// plugins/score-card/src/types/SystemScore.ts
/*
 * Copyright 2022 The Backstage Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
import { SystemScoreArea } from './SystemScoreArea';
import { ScoreSuccessEnum } from './SystemScoreSuccessEnum';

export interface SystemScore {
  systemEntityName: string;
  generatedDateTimeUtc: Date | string;
  scorePercent: number;
  scoreSuccess: ScoreSuccessEnum;
  scoringReviewer: string | undefined | null;
  scoringReviewDate: Date | string | undefined | null;
  areaScores: SystemScoreArea[];
}